Add UUID to 'temp' folder
The temporary folder gets deleted when orpheus_core_download finishes running.
However, the temporary folder will be used by any other instance of orpheusdl that is running and the deletion of this folder will interfere with other instances.

In order to prevent interference with other running instances of orpheusdl, here is a proposed change to create the temporary folder with a unique id that will not overlap.